requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
PyYAML==6.0
aws_cdk.aws_msk_alpha
cdk-nag
